requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
langchain
sentence-transformers
faiss-cpu